+++ .cspell.project-words.txt
@@ -42,3 +42,4 @@ dotenv
 autogen
 Notdroppedoff
 callsign
+vertipad
