+++ server/src/rest/api/mod.rs
@@ -1,0 +1,7 @@
+pub mod cancel;
+pub mod confirm;
+pub mod health;
+pub mod query;
+pub mod request;
+pub mod scan;
+pub mod utils;
