+++ server/src/rest/api/request.rs
@@ -1,0 +1,312 @@
+use super::utils::is_uuid;
+use crate::grpc::client::GrpcClients;
+use crate::rest_types::{FlightLeg, FlightRequest, Itinerary};
+use axum::{extract::Extension, Json};
+use chrono::Utc;
+use hyper::StatusCode;
+use lib_common::time::{datetime_to_timestamp, timestamp_to_datetime};
+
+//
+// Other Service Dependencies
+//
+use svc_pricing_client::pricing_grpc::{
+    pricing_request::ServiceType, PricingRequest, PricingRequests,
+};
+use svc_scheduler_client_grpc::grpc::{
+    Itinerary as SchedulerItinerary, QueryFlightPlan, QueryFlightRequest,
+};
+
+/// Don't allow excessively heavy loads
+const MAX_CARGO_WEIGHT_G: u32 = 1_000_000; // 1000 kg
+
+/// Parses the incoming flight plans for information the customer wants
+fn parse_flight(plan: &QueryFlightPlan) -> Option<FlightLeg> {
+    let Some(prost_time) = plan.estimated_departure.clone() else {
+        let error_msg = "no departure time in flight plan; discarding.";
+        rest_error!("(parse_flight) {}", &error_msg);
+        return None;
+    };
+
+    let Some(time_depart) = timestamp_to_datetime(&prost_time) else {
+        let error_msg = "can't convert prost timestamp to datetime.";
+        rest_error!("(parse_flight) {}", &error_msg);
+        return None;
+    };
+
+    let Some(prost_time) = plan.estimated_arrival.clone() else {
+        let error_msg = "(parse_flight) no arrival time in flight plan; discarding.";
+        rest_error!("(parse_flight) {}", &error_msg);
+        return None;
+    };
+
+    let Some(time_arrive) = timestamp_to_datetime(&prost_time) else {
+        let error_msg = "can't convert prost timestamp to datetime.";
+        rest_error!("(parse_flight) {}", &error_msg);
+        return None;
+    };
+
+    Some(FlightLeg {
+        flight_plan_id: plan.id.clone(),
+        vertiport_depart_id: plan.vertiport_depart_id.to_string(),
+        vertiport_arrive_id: plan.vertiport_arrive_id.to_string(),
+        timestamp_depart: time_depart,
+        timestamp_arrive: time_arrive,
+        distance_m: plan.estimated_distance as f32,
+        base_pricing: None,
+        currency_type: None,
+    })
+}
+
+// Get Available Flights
+///
+/// Search for available trips and return a list of [`Itinerary`].
+#[utoipa::path(
+    post,
+    path = "/cargo/request",
+    tag = "svc-cargo",
+    request_body = FlightRequest,
+    responses(
+        (status = 200, description = "List available flight plans", body = [Itinerary]),
+        (status = 400, description = "Request body is invalid format"),
+        (status = 500, description = "svc-scheduler or svc-pricing returned error"),
+        (status = 503, description = "Could not connect to other microservice dependencies")
+    )
+)]
+pub async fn request_flight(
+    Extension(mut grpc_clients): Extension<GrpcClients>,
+    Json(payload): Json<FlightRequest>,
+) -> Result<Json<Vec<Itinerary>>, StatusCode> {
+    rest_debug!("(query_flight) entry.");
+
+    //
+    // Validate Request
+    //
+
+    // Reject extreme weights
+    let weight_g: u32 = (payload.cargo_weight_kg * 1000.0) as u32;
+    if weight_g >= MAX_CARGO_WEIGHT_G {
+        let error_msg = format!("request cargo weight exceeds {MAX_CARGO_WEIGHT_G}.");
+        rest_error!("(query_flight) {}", &error_msg);
+        return Err(StatusCode::BAD_REQUEST);
+    }
+
+    // Check UUID validity
+    if !is_uuid(&payload.vertiport_arrive_id) {
+        let error_msg = "arrival port ID not UUID format.".to_string();
+        rest_error!("(query_flight) {}", &error_msg);
+        return Err(StatusCode::BAD_REQUEST);
+    }
+
+    if !is_uuid(&payload.vertiport_depart_id) {
+        let error_msg = "departure port ID not UUID format.".to_string();
+        rest_error!("(query_flight) {}", &error_msg);
+        return Err(StatusCode::BAD_REQUEST);
+    }
+
+    let mut flight_query = QueryFlightRequest {
+        is_cargo: true,
+        persons: None,
+        weight_grams: Some(weight_g),
+        vertiport_depart_id: payload.vertiport_depart_id,
+        vertiport_arrive_id: payload.vertiport_arrive_id,
+        earliest_departure_time: None,
+        latest_arrival_time: None,
+    };
+
+    let current_time = Utc::now();
+
+    // Time windows are properly specified
+    if let Some(window) = payload.time_arrive_window {
+        if window.timestamp_max <= current_time {
+            let error_msg = "max arrival time is in the past.".to_string();
+            rest_error!("(query_flight) {} {:?}", &error_msg, window.timestamp_max);
+            return Err(StatusCode::BAD_REQUEST);
+        }
+
+        let Some(ts) = datetime_to_timestamp(&window.timestamp_max) else {
+            let error_msg = "unable to convert datetime to timestamp.".to_string();
+            rest_error!("(query_flight) {} {:?}", &error_msg, window.timestamp_max);
+            return Err(StatusCode::BAD_REQUEST);
+        };
+
+        flight_query.latest_arrival_time = Some(ts);
+    }
+
+    if let Some(window) = payload.time_depart_window {
+        if window.timestamp_max <= current_time {
+            let error_msg = "max depart time is in the past.".to_string();
+            rest_error!("(query_flight) {} {:?}", &error_msg, window.timestamp_max);
+            return Err(StatusCode::BAD_REQUEST);
+        }
+
+        let Some(ts) = datetime_to_timestamp(&window.timestamp_max) else {
+            let error_msg = "unable to convert datetime to timestamp.".to_string();
+            rest_error!("(query_flight) {} {:?}", &error_msg, window.timestamp_max);
+            return Err(StatusCode::BAD_REQUEST);
+        };
+
+        flight_query.earliest_departure_time = Some(ts);
+    }
+
+    if flight_query.earliest_departure_time.is_none() && flight_query.latest_arrival_time.is_none()
+    {
+        let error_msg = "invalid time window.".to_string();
+        rest_error!("(query_flight) {}", &error_msg);
+        return Err(StatusCode::BAD_REQUEST);
+    }
+
+    //
+    // GRPC Request
+    //
+
+    let Some(mut scheduler_client) = grpc_clients.scheduler.get_client().await else {
+        let error_msg = "svc-scheduler unavailable.".to_string();
+        rest_error!("(query_flight) {}", &error_msg);
+        return Err(StatusCode::SERVICE_UNAVAILABLE);
+    };
+
+    let Some(mut pricing_client) = grpc_clients.pricing.get_client().await else {
+        let error_msg = "svc-pricing unavailable.".to_string();
+        rest_error!("(query_flight) {}", &error_msg);
+        return Err(StatusCode::SERVICE_UNAVAILABLE);
+    };
+
+    let request = tonic::Request::new(flight_query);
+    let response = scheduler_client.query_flight(request).await;
+    let Ok(response) = response else {
+        let error_msg = "svc-scheduler error.".to_string();
+        rest_error!("(query_flight) {} {:?}", &error_msg, response.unwrap_err());
+        rest_error!("(query_flight) invalidating svc-scheduler client.");
+        grpc_clients.scheduler.invalidate().await;
+        return Err(StatusCode::INTERNAL_SERVER_ERROR);
+    };
+
+    let itineraries: Vec<SchedulerItinerary> = response.into_inner().itineraries;
+
+    //
+    // Unpack flight itineraries
+    //
+
+    // List of lists of flights
+    let mut offerings: Vec<Itinerary> = vec![];
+    for itinerary in &itineraries {
+        let mut legs: Vec<FlightLeg> = vec![];
+
+        if let Some(plan) = &itinerary.flight_plan {
+            if let Some(leg) = parse_flight(plan) {
+                legs.push(leg);
+            }
+        }
+
+        for dh in &itinerary.deadhead_flight_plans {
+            if let Some(leg) = parse_flight(dh) {
+                legs.push(leg);
+            }
+        }
+
+        offerings.push(Itinerary {
+            id: itinerary.id.clone(),
+            legs,
+            base_pricing: None,
+            currency_type: Some("usd".to_string()),
+        })
+    }
+    rest_info!("(query_flight) found {} flight options.", offerings.len());
+
+    //
+    // Get pricing for each itinerary
+    //
+
+    // StatusUpdate message to customer?
+    // e.g. Got your flights! Calculating prices...
+    for mut itinerary in &mut offerings {
+        let mut pricing_requests = PricingRequests { requests: vec![] };
+
+        for leg in &itinerary.legs {
+            let pricing_query = PricingRequest {
+                service_type: ServiceType::Cargo as i32,
+                distance_km: leg.distance_m / 1000.0,
+                weight_kg: payload.cargo_weight_kg,
+            };
+
+            pricing_requests.requests.push(pricing_query);
+        }
+
+        // Make request, process response
+        let request = tonic::Request::new(pricing_requests);
+        let response = pricing_client.get_pricing(request).await;
+
+        let Ok(response) = response else {
+            let error_msg = "svc-pricing error.".to_string();
+            rest_error!("(query_flight) {} {:?}", &error_msg, response.unwrap_err());
+            rest_error!("(query_flight) invalidating svc-pricing client.");
+            grpc_clients.pricing.invalidate().await;
+            return Err(StatusCode::INTERNAL_SERVER_ERROR);
+        };
+
+        let response = response.into_inner();
+
+        for (price, mut leg) in response.prices.iter().zip(itinerary.legs.iter_mut()) {
+            leg.base_pricing = Some(*price);
+            leg.currency_type = Some("usd".to_string());
+        }
+
+        itinerary.base_pricing = Some(response.prices.iter().sum());
+    }
+
+    rest_debug!("(query_flight) exit with {} itineraries.", offerings.len());
+    Ok(Json(offerings))
+}
+
+#[cfg(test)]
+mod tests {
+    // Note this useful idiom: importing names from outer (for mod tests) scope.
+    use super::*;
+
+    #[test]
+    fn ut_parse_flight() {
+        let depart_time = datetime_to_timestamp(&Utc::now());
+        assert!(depart_time.is_some());
+
+        let fp = QueryFlightPlan {
+            id: "".to_string(),
+            pilot_id: "".to_string(),
+            vehicle_id: "".to_string(),
+            cargo: [123].to_vec(),
+            weather_conditions: "Sunny, no wind :)".to_string(),
+            vertiport_depart_id: "".to_string(),
+            pad_depart_id: "".to_string(),
+            vertiport_arrive_id: "".to_string(),
+            pad_arrive_id: "".to_string(),
+            estimated_departure: depart_time.clone(),
+            estimated_arrival: depart_time,
+            actual_departure: None,
+            actual_arrival: None,
+            flight_release_approval: None,
+            flight_plan_submitted: None,
+            flight_status: 0,
+            flight_priority: 0,
+            estimated_distance: 1000,
+        };
+
+        let Some(flight_plan) = parse_flight(&fp) else {
+            panic!();
+        };
+        assert_eq!(fp.id, flight_plan.flight_plan_id);
+        assert_eq!(fp.vertiport_depart_id, flight_plan.vertiport_depart_id);
+        assert_eq!(fp.vertiport_arrive_id, flight_plan.vertiport_arrive_id);
+
+        // Bad time arguments
+        {
+            let mut fp2 = fp.clone();
+            fp2.estimated_departure = None;
+            assert!(parse_flight(&fp2).is_none());
+        }
+
+        {
+            let mut fp2 = fp.clone();
+            fp2.estimated_arrival = None;
+            assert!(parse_flight(&fp2).is_none());
+        }
+    }
+}
